api_yamdb/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.0.4
psycopg2-binary==2.9
PyJWT==2.1.0
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
python-dotenv==0.21.0
requests==2.26.0
sqlparse==0.3.1
pytz==2020.1
python-decouple==3.7